@@ -1,68 +1,44 @@
 {
-  "name": "e-commerce",
-  "version": "0.0.1",
+  "name": "front",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "dev": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "generate": "bun scripts/openapi/ng-openapi-gen.ts"
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
     "@abacritt/angularx-social-login": "^2.2.0",
-    "@amcharts/amcharts5": "^5.6.2",
-    "@amcharts/amcharts5-geodata": "^5.1.1",
-    "@angular/animations": "^17.0.5",
-    "@angular/common": "^17.0.5",
-    "@angular/compiler": "^17.0.5",
-    "@angular/core": "^17.0.5",
-    "@angular/forms": "^17.0.5",
-    "@angular/google-maps": "^17.0.2",
-    "@angular/platform-browser": "^17.0.5",
-    "@angular/platform-browser-dynamic": "^17.0.5",
-    "@angular/router": "^17.0.5",
-    "@asymmetrik/ngx-leaflet": "^17.0.0",
-    "@lordicon/element": "^1.4.2",
-    "angular-leaflet": "^0.0.17",
-    "bootstrap": "5.3.2",
-    "jquery": "^3.7.1",
-    "leaflet": "^1.9.4",
-    "lord-icon-element": "^4.1.2",
-    "lottie-web": "^5.12.2",
-    "masonry-layout": "^4.2.2",
-    "ng-otp-input": "^1.9.3",
-    "ngx-cookie-service": "^17.0.0",
-    "ngx-masonry": "^14.0.1",
-    "ngx-slick-carousel": "^17.0.0",
-    "ngx-toastr": "^18.0.0",
+    "@angular/animations": "^17.3.0",
+    "@angular/common": "^17.3.0",
+    "@angular/compiler": "^17.3.0",
+    "@angular/core": "^17.3.0",
+    "@angular/forms": "^17.3.0",
+    "@angular/platform-browser": "^17.3.0",
+    "@angular/platform-browser-dynamic": "^17.3.0",
+    "@angular/router": "^17.3.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
     "primeflex": "^3.3.1",
-    "primeicons": "^6.0.1",
-    "primeng": "^17.0.0-rc.1",
-    "rxjs": "~7.8.1",
-    "simplebar-angular": "^3.2.4",
-    "slick-carousel": "^1.8.1",
-    "tslib": "^2.6.2",
-    "zone.js": "~0.14.2"
+    "primeicons": "^7.0.0",
+    "primeng": "^17.16.1",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.14.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.5",
-    "@angular/cli": "~17.0.5",
-    "@angular/compiler-cli": "^17.0.5",
-    "@types/jasmine": "~5.1.4",
-    "@types/leaflet": "^1.9.8",
-    "jasmine-core": "~5.1.1",
-    "json-schema-ref-parser": "^9.0.9",
-    "karma": "~6.4.2",
+    "@angular-devkit/build-angular": "^17.3.6",
+    "@angular/cli": "^17.3.6",
+    "@angular/compiler-cli": "^17.3.0",
+    "@types/jasmine": "~5.1.0",
+    "jasmine-core": "~5.1.0",
+    "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.1",
+    "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-openapi-gen": "^0.51.0",
-    "typescript": "~5.2.2"
-  },
-  "overrides": {
-    "autoprefixer": "10.4.5"
+    "typescript": "~5.4.2"
   }
 }
