@@ -1,5 +1,5 @@
 import { NgModule } from "@angular/core";
-import { RouterModule, Routes } from "@angular/router";
+import { RouterModule, type Routes } from "@angular/router";
 
 const routes: Routes = [
   { path: "auth", loadChildren: () => import("./auth/auth.module").then((m) => m.AuthModule) },
