@@ -1,32 +1,63 @@
-import { GoogleLoginProvider, GoogleSigninButtonModule, SocialAuthServiceConfig, SocialLoginModule } from '@abacritt/angularx-social-login';
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { environment } from 'src/environments/environment';
-import { FormModule } from '../shared/bootstrap/forms/form.module';
-import { AuthRoutingModule } from './auth.routing.module';
-import { AuthLayoutComponent } from './components/auth-layout/auth-layout.component';
-import { LoginComponent } from './components/login/login.component';
-import { RegisterComponent } from './components/register/register.component';
+import {
+  GoogleLoginProvider,
+  GoogleSigninButtonModule,
+  SocialLoginModule,
+  type SocialAuthServiceConfig,
+} from "@abacritt/angularx-social-login";
+import { CommonModule } from "@angular/common";
+import { NgModule } from "@angular/core";
+import { FormsModule, ReactiveFormsModule } from "@angular/forms";
+import { TranslateModule } from "@ngx-translate/core";
+import { ButtonModule } from 'primeng/button';
+import { CheckboxModule } from 'primeng/checkbox';
+import { DividerModule } from 'primeng/divider';
+import { ImageModule } from 'primeng/image';
+import { InputOtpModule } from 'primeng/inputotp';
+import { InputTextModule } from 'primeng/inputtext';
+import { PasswordModule } from 'primeng/password';
+import { ProgressSpinnerModule } from 'primeng/progressspinner';
+import { AuthRoutingModule } from "./auth.routing.module";
+import { AuthLayoutComponent } from "./components/auth-layout/auth-layout.component";
+import { ForgotPasswordComponent } from "./components/forgot-password/forgot-password.component";
+import { LoginComponent } from "./components/login/login.component";
+import { OtpComponent } from "./components/otp/otp.component";
+import { RegisterComponent } from "./components/register/register.component";
+import { ResetPasswordComponent } from "./components/reset-password/reset-password.component";
 
 const declarations = [
   AuthLayoutComponent,
+  ForgotPasswordComponent,
   LoginComponent,
+  OtpComponent,
   RegisterComponent,
+  ResetPasswordComponent,
+];
+
+const primengImports = [
+  ButtonModule,
+  CheckboxModule,
+  DividerModule,
+  ImageModule,
+  InputOtpModule,
+  InputTextModule,
+  PasswordModule,
+  ProgressSpinnerModule,
 ];
 
 const imported = [
+  TranslateModule,
   CommonModule,
-  FormModule,
+  FormsModule,
+  ReactiveFormsModule,
   GoogleSigninButtonModule,
   SocialLoginModule,
   AuthRoutingModule,
 ];
 
-
-// TODO: add client id to environment
 @NgModule({
   imports: [
-    ...imported
+    ...imported,
+    ...primengImports,
   ],
   providers: [
     {
@@ -37,7 +68,7 @@ const imported = [
           {
             id: GoogleLoginProvider.PROVIDER_ID,
             provider: new GoogleLoginProvider(
-              environment.google.clientId
+              '310931651035-k3gahf31rm8q3k4f9pft4s5k99r3g284.apps.googleusercontent.com'
             )
           },
         ],
@@ -47,6 +78,8 @@ const imported = [
       } as SocialAuthServiceConfig,
     }
   ],
-  declarations: [...declarations]
+  declarations: [
+    ...declarations,
+  ],
 })
 export class AuthModule { }
