@@ -1,8 +1,11 @@
 import { NgModule } from "@angular/core";
-import { RouterModule, Routes } from "@angular/router";
+import { RouterModule, type Routes } from "@angular/router";
 import { AuthLayoutComponent } from "./components/auth-layout/auth-layout.component";
+import { ForgotPasswordComponent } from "./components/forgot-password/forgot-password.component";
 import { LoginComponent } from "./components/login/login.component";
+import { OtpComponent } from "./components/otp/otp.component";
 import { RegisterComponent } from "./components/register/register.component";
+import { ResetPasswordComponent } from "./components/reset-password/reset-password.component";
 
 const routes: Routes = [
   {
@@ -10,13 +13,25 @@ const routes: Routes = [
     component: AuthLayoutComponent,
     children: [
       {
+        path: 'register',
+        component: RegisterComponent,
+      },
+      {
         path: 'login',
         component: LoginComponent
       },
       {
-        path: 'register',
-        component: RegisterComponent
+        path: 'otp',
+        component: OtpComponent
       },
+      {
+        path: 'forgot-password',
+        component: ForgotPasswordComponent,
+      },
+      {
+        path: 'reset-password',
+        component: ResetPasswordComponent,
+      }
     ]
   },
 ];
