@@ -1,7 +1,9 @@
-import { Component } from "@angular/core";
+import { Component, ViewEncapsulation } from "@angular/core";
 
 @Component({
   selector: 'app-auth-layout',
   templateUrl: './auth-layout.component.html',
+  styleUrls: ['./auth-layout.component.scss'],
+  encapsulation: ViewEncapsulation.None,
 })
 export class AuthLayoutComponent { }
