@@ -1,17 +1,16 @@
 import { GoogleLoginProvider, SocialAuthService } from "@abacritt/angularx-social-login";
 import { Component, type OnInit } from "@angular/core";
 import { FormControl, FormGroup, Validators } from "@angular/forms";
-import { EmailValidator } from '@shared/validators/email.validator';
+
+// Bloquer le compte après 5 tentatives de connexion
 
 @Component({
-  selector: "app-auth-register",
-  templateUrl: "./register.component.html",
-  styleUrls: ["./register.component.scss"],
+  selector: "app-auth-otp",
+  templateUrl: "./otp.component.html",
 })
-export class RegisterComponent implements OnInit {
+export class OtpComponent implements OnInit {
   protected form!: FormGroup;
   protected isLoading!: boolean;
-  protected passwordMinLength = 8;
   constructor(private authService: SocialAuthService) { }
 
   public ngOnInit(): void {
@@ -28,12 +27,7 @@ export class RegisterComponent implements OnInit {
   private initForm() {
     this.form = new FormGroup({
       username: new FormControl('', [Validators.required]),
-      firstname: new FormControl('', [Validators.required]),
-      lastname: new FormControl('', [Validators.required]),
-      email: new FormControl('', [Validators.required, Validators.pattern(EmailValidator.pattern)]),
-      password: new FormControl('', [Validators.required, Validators.minLength(6)]),
-      hasReadContracts: new FormControl(false, Validators.requiredTrue),
-      enablePromotionalMails: new FormControl(false),
+      password: new FormControl('', [Validators.required]),
     });
   }
 
