@@ -1,16 +1,16 @@
 import { GoogleLoginProvider, SocialAuthService } from "@abacritt/angularx-social-login";
 import { Component, type OnInit } from "@angular/core";
 import { FormControl, FormGroup, Validators } from "@angular/forms";
-import { EmailValidator } from '@shared/validators/email.validator';
+
+// Mettre en place un token temporaire qui sera dans la query de l'url pour rénitiliser le mot de passe
+// Ce token sera obligatoire pour que le back end puisse réinitialiser le mot de passe
 
 @Component({
-  selector: "app-auth-register",
-  templateUrl: "./register.component.html",
-  styleUrls: ["./register.component.scss"],
+  selector: "app-auth-reset-password",
+  templateUrl: "./reset-password.component.html",
 })
-export class RegisterComponent implements OnInit {
+export class ResetPasswordComponent implements OnInit {
   protected form!: FormGroup;
-  protected isLoading!: boolean;
   protected passwordMinLength = 8;
   constructor(private authService: SocialAuthService) { }
 
@@ -27,13 +27,7 @@ export class RegisterComponent implements OnInit {
 
   private initForm() {
     this.form = new FormGroup({
-      username: new FormControl('', [Validators.required]),
-      firstname: new FormControl('', [Validators.required]),
-      lastname: new FormControl('', [Validators.required]),
-      email: new FormControl('', [Validators.required, Validators.pattern(EmailValidator.pattern)]),
-      password: new FormControl('', [Validators.required, Validators.minLength(6)]),
-      hasReadContracts: new FormControl(false, Validators.requiredTrue),
-      enablePromotionalMails: new FormControl(false),
+      password: new FormControl('', [Validators.required, Validators.minLength(this.passwordMinLength)]),
     });
   }
 
